Move Job & JobRun regional nav into containers directory
- They should live in the `containers` directory because they are
  connected to the redux store
- The 'components` directory is intended for components that are
  isolated from parent hierachy and setup
