+++ operator_ui/__tests__/containers/JobRuns/Show/RegionalNav.test.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import RegionalNav from 'components/JobRuns/RegionalNav'
+import RegionalNav from 'containers/JobRuns/Show/RegionalNav'
 import mountWithTheme from 'test-helpers/mountWithTheme'
 
 const CREATED_AT = '2019-06-11T14:37:42.077995-07:00'
 
-describe('components/JobRuns/RegionalNav', () => {
-  it('displays an overview & json tab', () => {
+describe('containers/JobRuns/Show/RegionalNav', () => {
+  it('displays an overview & json tab by default', () => {
     const component = mountWithTheme(<RegionalNav />)
 
     expect(component.text()).toContain('Overview')
@@ -13,7 +13,7 @@ describe('components/JobRuns/RegionalNav', () => {
     expect(component.text()).not.toContain('Error Log')
   })
 
-  it('displays an error log tab when the status is "errored"', () => {
+  it('also displays an error log tab when the status is "errored"', () => {
     const jobRun = { status: 'errored', createdAt: CREATED_AT }
     const component = mountWithTheme(<RegionalNav jobRun={jobRun} />)
