+++ operator_ui/src/containers/Jobs/Definition.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import { connect } from 'react-redux'
 import { withStyles } from '@material-ui/core/styles'
 import Grid from '@material-ui/core/Grid'
 import Card from '@material-ui/core/Card'
@@ -6,9 +7,8 @@ import CardContent from '@material-ui/core/CardContent'
 import Typography from '@material-ui/core/Typography'
 import Divider from '@material-ui/core/Divider'
 import Content from 'components/Content'
-import RegionalNav from 'components/Jobs/RegionalNav'
 import PrettyJson from 'components/PrettyJson'
-import { connect } from 'react-redux'
+import RegionalNav from './RegionalNav'
 import { fetchJob, createJobRun } from 'actions'
 import jobSelector from 'selectors/job'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
