+++ operator_ui/src/containers/Jobs/Show.tsx
@@ -7,7 +7,7 @@ import KeyValueList from '@chainlink/styleguide/components/KeyValueList'
 import CardTitle from '@chainlink/styleguide/components/Cards/Title'
 import JobRunsList from '../../components/JobRuns/List'
 import Content from '../../components/Content'
-import RegionalNav from '../../components/Jobs/RegionalNav'
+import RegionalNav from './RegionalNav'
 import { fetchJob, fetchJobRuns } from '../../actions'
 import jobSelector from '../../selectors/job'
 import jobRunsByJobIdSelector from '../../selectors/jobRunsByJobId'
